superio/smsc/sio1036: Support 16-bit IO port addressing
SMSC/Microchip 1036 can be strapped to 4E/4D and 164E/164D so make 
source code support 16 bits addressing.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/47647
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/superio/smsc/sio1036/sio1036_early_init.c
@@ -10,13 +10,13 @@
 
 static inline void sio1036_enter_conf_state(pnp_devfn_t dev)
 {
-	u8 port = dev >> 8;
+	u16 port = dev >> 8;
 	outb(0x55, port);
 }
 
 static inline void sio1036_exit_conf_state(pnp_devfn_t dev)
 {
-	u8 port = dev >> 8;
+	u16 port = dev >> 8;
 	outb(0xaa, port);
 }
